+++ example/lib/main.dart
@@ -3,6 +3,7 @@
 ///
 ///
 ///
+
 import 'src/view.dart' show MyApp, runApp;
 
 void main() => runApp(MyApp());